moved validator middleware to it's own package and added test suite
