+++ pkg/grpc/middleware/validator/validator_wrapper.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package middleware
+package validator
 
 import (
 	"context"
